rename test dir name
